refactor(threads): move pthread_join to its own module;
